requirements.txt
dependency-injector
flask
pygithub
pyyaml